move the operation state concept into its own header
